Difficulty and location setup
Set up the difficulty methods to create a specific number of text files at random locations, each having their own message so the player can track how many files they found. A future commit will include hints to be printed to the terminal for easy difficulty.

Furthermore, created an array containing all the possible locations (limited to 10 for efficiency) that could be chosen at random for a player to find a text file in.