Desktop and download organisation
Set up the cleanup method using a userChoice parameter which executes the code to clean-up the users desktop or the download folder depending on their choice in the menu.  Additionally, added in a check to create downloaded documents/images if they do not exist as this can help the user distinguish which files are from the desktop or the downloads folder.

Furthermore, added a location select method which allows the user to choose a location of their own to be cleaned up.