Delete user and admin right check
Implemented a way for the user to remove others users and their profile to mimic wiping their account from the system but also added a safeguard to avoid the user from deleting themself.

Furthermore, added a simple admin check within the menu to ensure that the user running the script has admin privileges.