Clean-up separation and FolderBrowserDialog
Added in separated clean-up functions to avoid any overlap of cleaning other destinations at the same time (e.g. the program organises the desktop and downloads although the user chose a location on their E: drive.)

Also implemented a FolderBrowserDialog object to allow the process of the user choosing their folder to be easier rather than inputting the full path.

Lastly, organised variables such as the downloaded images/documents locations to be created within the functions and applied the same method of creating the folders within the users selected location to easily find them.